--- codex-ui/src/vue/form-field/index.ts
@@ -1,3 +1,0 @@
-import FormField from './FormField.vue';
-
-export { FormField };
