--- codex-ui/src/vue/form-section/index.ts
@@ -1,4 +1,0 @@
-export * from './FormSection.types.js';
-import FormSection from './FormSection.vue';
-
-export { FormSection };
