+++ codex-ui/src/vue/index.ts
@@ -1,6 +1,5 @@
 export * from './button';
-export * from './form-section';
-export * from './form-field';
+export * from './form';
 export * from './heading';
 export * from './input';
 export * from './row';
